backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.13.2
flasgger
Flask==3.1.2
flask-cors
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.7
langchain-core==1.2.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
MarkupSafe==3.0.3
numpy==2.4.1
openai==2.16.0
orjson==3.11.6
ormsgpack==1.12.2
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
supabase
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
Werkzeug==3.1.5
xxhash==3.6.0
zstandard==0.25.0
googlemaps
